# Optimize JMT and RANDOM-NORMAL for SBCL